refactor(item): update item entity and test item
