+++ src/wrapper/entities/ItemFactory.ts
@@ -1,26 +1,21 @@
-import { ItemSprites } from '@/renderer/canvas/sprite/entities/ItemSprites';
+import { ItemAnimation } from '@/renderer/canvas/sprite/entities/ItemAnimation';
+import { EntityAnimation } from '@/renderer/canvas/sprite/EntityAnimation';
 import { SpriteResource } from '@/renderer/canvas/sprite/SpriteResource';
-import { EntityState } from '@/wrapper/entities/Entity';
-import { ItemEntity } from './ItemEntity';
+import { EntityState } from '@/wrapper/entities/EntityState';
+import { ItemEntity } from '@/wrapper/entities/ItemEntity';
 
 export abstract class ItemFactory implements ItemEntity {
     
-    protected readonly sprites: ItemSprites;
-
-    protected x: number;
-    protected y: number;
+    //protected readonly sprites: ItemSprites;
+    protected animation: ItemAnimation | undefined;
+    protected x: number = 0;
+    protected y: number = 0;
     protected velocity: number = 1.0;
-    protected health: number = 1;
+    protected health: number = 30;
     protected state: EntityState = EntityState.IDLE;
 
-    public constructor(sprites: ItemSprites, x: number, y: number) {
-        this.sprites = sprites;
-        this.x = x;
-        this.y = y;
-    }
-
-    public async load(): Promise<void> {
-        await this.sprites.loadAllSprites();
+    public setAnimation(animation: EntityAnimation): void {
+        this.animation = animation;
     }
 
     public getX(): number {
@@ -39,36 +34,55 @@ export abstract class ItemFactory implements ItemEntity {
         this.y = y;
     }
 
-    public setVelocity(velocity: number): void {
-        this.velocity = velocity;
-    }
-
     public getVelocity(): number {
         return this.velocity;        
     }
 
-    public getCurrentSprite(): SpriteResource {
-        return this.sprites.getCurrentSprite();
-    }
-
-    public setCurrentState(state: EntityState): void {
-        this.state = state;
-        this.sprites.setCurrentSprite(state);
-    }
-
-    public setCurrentTemporaryState(state: EntityState, afterState: EntityState): void {
-        this.state = state;
-        this.sprites.setCurrentSprite(state, () => {
-            this.setCurrentState(afterState);
-        });
+    public setVelocity(velocity: number): void {
+        this.velocity = velocity;
     }
 
     public getCurrentState(): EntityState {
         return this.state;
     }
 
+    public setCurrentState(state: EntityState): void {
+        if(!this.animation) throw new Error(`${this.constructor.name} doesn't have an animation`)
+        
+        this.state = state;
+        this.animation.setCurrentSprite(state);
+    }
+
+    public getCurrentSprite(): SpriteResource {
+        if(!this.animation) throw new Error(`${this.constructor.name} doesn't have an animation`);
+
+        return this.animation.getCurrentSprite();
+    }
+
+    public setCurrentTemporaryState(state: EntityState, afterState: EntityState): void {
+        if(!this.animation) throw new Error(`${this.constructor.name} doesn't have an animation`)
+        
+        this.state = state;
+        this.animation.setCurrentSprite(state, () => {
+            this.setCurrentState(afterState);
+        });
+    }
+
+    public getAnimation(): EntityAnimation {
+        if(!this.animation) throw new Error(`${this.constructor.name} doesn't have an animation`);
+        
+        return this.animation;
+    }
+
+    public isIdle(): boolean {
+        return this.state == EntityState.IDLE;
+    }
+
     public abstract idle(): void;
+
     public abstract collect(): void;
+
     public abstract effect(): void;
-    public abstract destroy(): void;
+
+    public abstract expire(): void;
 }
